Add policy to crud Job

+++ app/Http/Controllers/JobController.php
@@ -3,13 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
     // @desc Show all job listings
     // @route GET /jobs
     public function index() : View {
@@ -33,12 +36,17 @@ class JobController extends Controller
     // @desc Show edit job form
     // @route GET /jobs/edit/{$id}
     public function edit(Job $job) : View {
+        //Check if user authorize
+        $this->authorize('delete', $job);
         return view('jobs.edit')->with('job', $job);
     }
 
     // @desc Update job listing
     // @route PUT /jobs/{$id}
-    public function update(Request $request, Job $job){
+    public function update(Request $request, Job $job) : string {
+        //Check if user authorize
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -115,6 +123,9 @@ class JobController extends Controller
     // @desc Delete a job listing
     // @route DELETE /jobs/{$id}
     public function destroy(Job $job) : RedirectResponse {
+        //Check if user authorize
+        $this->authorize('delete', $job);
+
         //If logo, then delete it
         if($job->company_logo){
             Storage::delete('public/logos/' . $job->company_logo);
